requirements.txt
transformers
torch
torchvision
Pillow
opencv-python
openai-whisper
google-generativeai
tqdm
accelerate
